create translation to formal German
in German there are two ways of approaching someone. The existing
translation german.yaml uses the informal way ("du"). This adds the
formal way of approaching the user treating it as if it was a new
language: german_formal. Course authors can choose which one to use by
setting the preferred language option.